[medialibrary-185] Fix references to Organizer object class
