+++ src/media/tools/movies/namelist.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 #
-# Copyright 2023 Chris Josephes
+# Copyright 2024 Chris Josephes
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -30,7 +30,7 @@ Module to report on all named within a movie
 
 import os
 import argparse
-from media.generic.sorting.lists import Organizer
+from media.generic.sorting.organizer import Organizer
 from media.tools.common import load_movies
